# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backtest Platform Research API",
    "version": "0.6.0"
  },
  "paths": {
...
    "/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "description": "Paginated list of run summaries, one current row per run_id (latest wins).",
        "operationId": "list_runs_runs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page index",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page index"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "page size",
              "default": 50,
              "title": "Limit"
            },
            "description": "page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_RunSummary__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "description": "Trigger one IS run: build+validate a RunConfig, judge it, append to the ledger.\n\nA bad config (unknown preset, reversed window, ...) surfaces as 422 from the\ndomain validator rather than a 500.",
        "operationId": "create_run_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RunRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/runs/{run_id}/equity": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Run Equity",
        "description": "Equity curve + running drawdown for one run (from the series sidecar).",
        "operationId": "run_equity_runs__run_id__equity_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/runs/{run_id}/report": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Run Report",
        "description": "One-shot aggregate for the Run-Report page (verdict / segments / monthly /\ndrawdown events / cost sensitivity). 404 if the run is unknown; every\nunsourceable field is ``null`` (never fabricated).",
        "operationId": "run_report_runs__run_id__report_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RunReportData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/notebook": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Run Notebook",
        "description": "Download a prefilled ``.ipynb`` for one run (Open-in-notebook). 404 if unknown.\n\nReturns the notebook JSON with a ``Content-Disposition: attachment`` header so a\nbrowser saves ``run_<id>.ipynb``; the template loads the run from the repo stores\n(not the REST API) when the operator runs it in the repo venv Jupyter.",
        "operationId": "run_notebook_runs__run_id__notebook_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "description": "Structured error body shared by every failure response."
      },
...
      "Envelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "Envelope",
        "description": "The single response shape shared by every endpoint.\n\nGeneric over the ``data`` payload: ``response_model=Envelope[RunRow]`` types\nthe OpenAPI ``data`` schema (and the frontend's generated types) per endpoint.\nUnparametrised ``Envelope`` \u2261 ``Envelope[Any]`` \u2014 the ``ok``/``fail``/``pending``\nhelpers and any endpoint without a specific shape keep working untyped, with\nno field loss."
      },
...
      "Envelope_RunRecord_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "Envelope[RunRecord]"
      },
      "Envelope_RunReportData_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunReportData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "Envelope[RunReportData]"
      },
...
      "Envelope_list_RunSummary__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RunSummary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "Envelope[list[RunSummary]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunCreateRequest": {
        "properties": {
          "hypothesis": {
            "type": "string",
            "minLength": 1,
            "title": "Hypothesis",
            "description": "\u9810\u5148\u8a3b\u518a\uff1a\u9019\u500b run \u5728\u9a57\u4ec0\u9ebc"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy",
            "description": "Registered strategy name (see GET /strategies)"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "description": "Strategy params \u2014 validated at dispatch"
          },
          "stocks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Stocks",
            "description": "Stock ids to run"
          },
          "is_start": {
            "type": "string",
            "format": "date",
            "title": "Is Start"
          },
          "is_end": {
            "type": "string",
            "format": "date",
            "title": "Is End"
          },
          "engine": {
            "type": "string",
            "title": "Engine",
            "description": "sim (zipline/vectorbt removed 2026-07-03, ADR-037)",
            "default": "sim"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "hypothesis",
          "strategy",
          "stocks",
          "is_start",
          "is_end"
        ],
        "title": "RunCreateRequest",
        "description": "Trigger one IS run \u2014 the HTTP form of a ``RunConfig`` (validated downstream)."
      },
      "RunRecord": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunRecord",
        "description": "A full ledger record (``GET /runs/{id}`` / ``POST /runs``). Shape varies by\nrun; only the identity field is guaranteed, the rest pass through."
      },
      "RunReportData": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "verdict": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "segments": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segments"
          },
          "monthly_returns": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Returns"
          },
          "monthly_returns_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Returns Note"
          },
          "drawdown_events": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drawdown Events"
          },
          "cost_sensitivity": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Sensitivity"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunReportData",
        "description": "``GET /runs/{id}/report`` \u2014 one-shot Run-Report aggregate.\n\nBlocks are loosely typed (dict/list) on purpose: each carries an honest ``null``\nwhen its source is absent (verdict inputs, sidecar-derived charts), so a strict\nnested schema would misrepresent the contract. ``extra='allow'`` keeps every\nsub-field flowing to the frontend."
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          },
          "gate_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Status"
          },
          "hypothesis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hypothesis"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "is_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Start"
          },
          "is_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is End"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunSummary",
        "description": "One row of ``GET /runs`` (``_SUMMARY_KEYS`` projection)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}